2.3 Alignment CSS
In the last lesson you learned how to control the overall size of elements in your layout. In this lesson, we’ll be stepping through how to align your elements in the way that you want so you can do things like creating columns, positioning logos, adjusting text, and so on.

You’ll learn the classic method of centering in CSS, the basics of the float and clear properties, how to align text, and the difference between block and inline elements.